r Extract astring.c
